+++ core/src/main/java/com/github/starnowski/jamolingo/core/operators/apply/ApplySearchToMongoPipelineParser.java
@@ -1,0 +1,6 @@
+package com.github.starnowski.jamolingo.core.operators.apply;
+
+/** ApplySearchToMongoPipelineParser type. */
+public interface ApplySearchToMongoPipelineParser {
+  ApplyOperatorResult parse(SearchApplyItemContext context);
+}