+++ core/src/main/java/com/github/starnowski/jamolingo/core/operators/apply/IdentityItemParser.java
@@ -1,0 +1,13 @@
+package com.github.starnowski.jamolingo.core.operators.apply;
+
+import com.github.starnowski.jamolingo.core.api.EdmPropertyMongoPathResolver;
+import org.apache.olingo.server.api.uri.queryoption.ApplyItem;
+
+/** IdentityItemParser type. */
+public class IdentityItemParser implements ApplyItemParser {
+  @Override
+  public ApplyOperatorResult parse(
+      ApplyItem applyItem, EdmPropertyMongoPathResolver edmMongoContextFacade) {
+    return DefaultApplyOperatorResult.builder().build();
+  }
+}