+++ core/src/main/java/com/github/starnowski/jamolingo/core/operators/apply/JoinItemParser.java
@@ -1,0 +1,13 @@
+package com.github.starnowski.jamolingo.core.operators.apply;
+
+import com.github.starnowski.jamolingo.core.api.EdmPropertyMongoPathResolver;
+import org.apache.olingo.server.api.uri.queryoption.ApplyItem;
+
+/** JoinItemParser type. */
+public class JoinItemParser implements ApplyItemParser {
+  @Override
+  public ApplyOperatorResult parse(
+      ApplyItem applyItem, EdmPropertyMongoPathResolver edmMongoContextFacade) {
+    throw new UnsupportedOperationException("Join set transformation is not yet implemented.");
+  }
+}